Fix compatibility with ksh.
With older versions of ksh, running functions of ksh is not consistent.
So instead of doing it for setting up the pipenv environments, change
the directory back on each loop.

Using a sed for removing the PIPENV_* variables was not entirely
portable. The best way is to do it manually.

Doing unset on the function that the shell is currently executing
segfaults older versions of ksh.
So, instead, we should call deactivate with nondestructive in
pipenv_deactivate and unset it manually.

Finally, getting the error message from pipenv_activate was not
consistent. So ignore the assert for ksh.