Rework loading dotenv values and unload, and robustify tests.
Contrary to what I believed, Pipenv overwrites the environment values
when loading the dotenv file.

It is an issue for us because we will need to save and restore the
already existing shell variables on load and unload.

After robustifying the tests, I found out that `printf %b` is not 100%
compatible with python `json.dumps`.
In order to avoid having some difficult and fragile value escaping,
encode the values in base64.

Unfortunately, `base64` command is not POSIX, but part of GNU coreutils.
So, to make it portable, use python as we know that it is available for
Pipenv.