Get python executable from PATH instead of pipenv shebang.
Using the pipenv shebang to get the python executable guaranteed that we
use the same python interpreter for pipenv and for loading/unloading the
dotenv file.

Unfortunately, this doesn't work when pipenv is redefined as a shim like
with asdf.

The solution is to look for a python executable that has a pipenv module
available in the list of all python executable accessible according to
the PATH.