Fix finding Pipenv python executable for pipx
Pipx seems to install Pipenv in `~/.local/bin/pipenv` with a shebang
path to python designed to load the local user environment.

Unfortunately, this shebang path is a symlink and the PATH look up
introduced by 5a9c1461748a8f952 is not smart enough to find the correct
environment.

One solution is to look back at the shebang again first, and then do the
the PATH look up.

The shebang parsing regex is more complex to avoid incorrect shebang
issues.