Add 'PYVENV_ACTIVATE_TOP_LEVEL_ENV' option.
Use top-level environment instead of lowest level by default.
By default, we choose the lowest level of environment because this is
the one that should have the correct environment for the current
project.
Set it to '1' to activate using top-level environment.
Default is '0', which means that we use the lowest level of environment.

WARNING: Activating this option can have some unwanted behaviour with
Pipenv.
Pipenv commands always use the lowest level of environment even if a top
level environment is already activated.
So `pipenv run <command>` will not have the expected behaviour.